+++ front/src/components/Modal/NewsModal.jsx
